refactor: move api service code into bloc and make module structure leaner
