+++ lib/app/services.dart
@@ -1,4 +1,3 @@
-export 'services/api.dart';
 export 'services/authentication_storage.dart';
 export 'services/network.dart';
 export 'services/user.dart';
